requirements.txt
beautifulsoup4==4.9.3
Pillow==8.0.1
pytesseract==0.3.7
soupsieve==2.1
